+++ packages/terra-graphs/src/components/Bubble/BubbleMultipleDataset.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import '../Graph.module.scss';
-import './BubbleMultipleDataset.module.scss';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 
 const propTypes = {
   /**
@@ -26,8 +26,22 @@ const propTypes = {
 const BubbleMultipleDataset = ({
   graphConfig, dataset, graphID, timeout,
 }) => {
+  const [graph, setGraph] = React.useState();
+  const graphLoadedRef = React.useRef();
+  const skipreflowRef = React.useRef();
+
+  // creation of canvas
   React.useEffect(() => {
-    const graph = Carbon.api.graph(graphConfig);
+    if (!graph) {
+      setGraph(Carbon.api.graph(graphConfig));
+    }
+  }, [graph, graphConfig]);
+
+  // initial dataset load
+  React.useEffect(() => {
+    if (!graph || graphLoadedRef.current) {
+      return;
+    }
     const timeoutIds = [];
 
     if (dataset) {
@@ -39,6 +53,7 @@ const BubbleMultipleDataset = ({
               : ''),
             timeout[index],
           );
+
           timeoutIds.push(timeoutId);
         });
       } else {
@@ -47,10 +62,34 @@ const BubbleMultipleDataset = ({
         });
       }
     }
+    graphLoadedRef.current = true;
+
+    // eslint-disable-next-line consistent-return
     return () => {
       timeoutIds.forEach((id) => { clearTimeout(id); });
     };
-  }, [graphConfig, dataset, timeout]);
+  }, [graph, dataset, timeout]);
+
+  // panning
+  React.useEffect(() => {
+    if (!graph) {
+      return;
+    }
+    if (!skipreflowRef.current) {
+      skipreflowRef.current = true;
+      return;
+    }
+
+    graph.config.axis.x.upperLimit = graphConfig.axis.x.upperLimit;
+    graph.config.axis.x.lowerLimit = graphConfig.axis.x.lowerLimit;
+
+    const newDataset = {
+      panData: utils.deepClone(dataset.panData),
+      eventline: utils.deepClone(dataset.eventline),
+    };
+
+    graph.reflowMultipleDatasets(newDataset);
+  }, [graph, dataset, graphConfig]);
 
   return (
     <div id={`${graphID}-canvasContainer`}>