+++ packages/terra-graphs/CHANGELOG.md
@@ -14,7 +14,10 @@
   * Added reflow to the Line component.
   * Added CSS and new example tests for timeline graph.
   * Added reflow to the Paired result component.
+  * Added reflow to the Bar component.
+  * Added reflow to the Bubble component.
   * Added reflow to the Timeline component.
+
 
 * Changed
   * Made changes to WDIO test url's according to terra rules.